requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
av==10.0.0
beautifulsoup4==4.11.2
charset-normalizer==3.1.0
discord.py==2.2.2
dnspython==2.3.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openpyxl==3.1.2
pymongo==4.3.3
soupsieve==2.4
yarl==1.8.2
